requirements.txt
streamlit==1.29.0
tensorflow==2.15.0
numpy==1.21.5
opencv-python==4.9.0.80
Pillow==10.2.0
tensorflow-hub==0.16.1
